requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Object Detection
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Face Detection
facenet-pytorch>=2.5.3

# Web App
streamlit>=1.28.0

# Training & Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
